feat: 해외주식(미국) 모의투자 지원 + market 파라미터로 시장별 시작
- 해외주식 KIS API 모듈 신규 (overseas_market.py, overseas_order.py)
- 미국 장 캘린더 추가 (NYSE 휴일, ET 시간대, 09:30~16:00)
- 가격 타입 int→float 통일 (미국 소수점 가격 지원)
- MarketType enum + detect_market_type()으로 심볼 자동 감지
- 엔진 국내/해외 듀얼 라우팅 (시세, 주문, 잔고 분기)
- /trading/start에 market 파라미터 추가 ("domestic"/"us"/null)
- 시장별 독립 장시간 체크 (국내만/미국만/둘다)
- 백테스트 US 티커 지원 (yfinance 패스스루)
- 테스트 264개 (신규 74개: 해외시세 13, 해외주문 12, US캘린더 20, 엔진 20, 백테스트 9)

+++ app/kis/market.py
@@ -45,13 +45,13 @@ class KISMarketClient:
         return StockPrice(
             symbol=symbol,
             name=output.get("hts_kor_isnm", ""),
-            current_price=int(output.get("stck_prpr", 0)),
-            change=int(output.get("prdy_vrss", 0)),
+            current_price=float(output.get("stck_prpr", 0)),
+            change=float(output.get("prdy_vrss", 0)),
             change_rate=float(output.get("prdy_ctrt", 0)),
             volume=int(output.get("acml_vol", 0)),
-            high=int(output.get("stck_hgpr", 0)),
-            low=int(output.get("stck_lwpr", 0)),
-            opening=int(output.get("stck_oprc", 0)),
+            high=float(output.get("stck_hgpr", 0)),
+            low=float(output.get("stck_lwpr", 0)),
+            opening=float(output.get("stck_oprc", 0)),
         )
 
     async def get_daily_chart(
@@ -91,10 +91,10 @@ class KISMarketClient:
             chart.append(
                 ChartData(
                     date=date,
-                    open=int(item.get("stck_oprc", 0)),
-                    high=int(item.get("stck_hgpr", 0)),
-                    low=int(item.get("stck_lwpr", 0)),
-                    close=int(item.get("stck_clpr", 0)),
+                    open=float(item.get("stck_oprc", 0)),
+                    high=float(item.get("stck_hgpr", 0)),
+                    low=float(item.get("stck_lwpr", 0)),
+                    close=float(item.get("stck_clpr", 0)),
                     volume=int(item.get("acml_vol", 0)),
                 )
             )
@@ -143,10 +143,10 @@ class KISMarketClient:
                 chart.append(
                     ChartData(
                         date=hour,
-                        open=int(item.get("stck_oprc", 0)),
-                        high=int(item.get("stck_hgpr", 0)),
-                        low=int(item.get("stck_lwpr", 0)),
-                        close=int(item.get("stck_prpr", 0)),
+                        open=float(item.get("stck_oprc", 0)),
+                        high=float(item.get("stck_hgpr", 0)),
+                        low=float(item.get("stck_lwpr", 0)),
+                        close=float(item.get("stck_prpr", 0)),
                         volume=int(item.get("cntg_vol", 0)),
                     )
                 )
